Added improved_hac to hac which is a much more efficient version of hac. It produces slightly different results which I'm almost positive is due to breaking ties differently. It is much faster though. To cluster 2 articles worth of spin groups until no two clusters have similarity > .66 it takes about 4 seconds as opposed to 30 seconds.
Also added an option in HMM constructor for how many articles to deal with. This is useful in debugging when we only want to do it on a small set of data

Added a decorator function in util to time function calls. Just decorate an function you want to time with @time_function and it will print out hte time every call to that function takes